Add NRT from YouVersion and Visiobible Refs #14
