+++ res/Windows/International Bible Society (IBS)/rvmetadata.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8" standalone="no"?>
+<RVBibleMetdata>
+    <name>(IBS) International Bible Society</name>
+    <abbreviation>ACV</abbreviation>
+    <displayAbbreviation>IBS</displayAbbreviation>
+    <version>1</version>
+    <revision>1</revision>
+    <licenseType>0</licenseType>
+    <license>Public Domain</license>
+</RVBibleMetdata>
