--- res/Windows/New Russian Translation/rvmetadata.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8" standalone="no"?>
-<RVBibleMetdata>
-    <name>New Russian Translation</name>
-    <abbreviation>ACV</abbreviation>
-    <displayAbbreviation>NRT</displayAbbreviation>
-    <version>1</version>
-    <revision>1</revision>
-    <licenseType>0</licenseType>
-    <license>Public Domain</license>
-</RVBibleMetdata>
