+++ res/Windows/Russian Synodal Translation (SYN)/rvmetadata.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8" standalone="no"?>
+<RVBibleMetdata>
+    <name>(NTC) Russian Synodal Translation</name>
+    <abbreviation>RST</abbreviation>
+    <version>1</version>
+    <revision>1</revision>
+    <licenseType>0</licenseType>
+    <license>Public Domain</license>
+</RVBibleMetdata>
