+++ src/BibHelp/BibleMap.py
@@ -1,5 +1,6 @@
 class BookName:
 
+    # TODO: Put information about the Testament in this struct
     def __init__(self, aliases: list, ntc_ru_long: str, en_abbr: str):
         self.aliases = aliases
         self.ntc_ru_long = ntc_ru_long
@@ -55,13 +56,14 @@ class BookNameManager:
         BookName(['Иисус Навин', 'Книга Иисуса Навина'], 'Иисус Навин', 'JOS'),
         BookName(['Судьи', 'Книга Судей израилевых'], 'Судьи', 'JDG'),
         BookName(['Руфь', 'Книга Руфи'], 'Руфь', 'RUT'),
-        BookName(['1 Царств', '1-я Царств', 'Первая книга Царств [Первая Самуила]'], '1 Царств',
+        BookName(['1 Царств', '1-я Царств', 'Первая книга Царств [Первая Самуила]', 'Первая книга Царств'], '1 Царств',
                  '1SA'),
-        BookName(['2 Царств', '2-я Царств', 'Вторая книга Царств [Вторая Самуила]'], '2 Царств',
+        BookName(['2 Царств', '2-я Царств', 'Вторая книга Царств [Вторая Самуила]', 'Вторая книга Царств'], '2 Царств',
                  '2SA'),
-        BookName(['3 Царств', '3-я Царств', 'Третья книга Царств [Первая Царей]'], '3 Царств',
+        BookName(['3 Царств', '3-я Царств', 'Третья книга Царств [Первая Царей]', 'Третья книга Царств'], '3 Царств',
                  '1KI'),
-        BookName(['4 Царств', '4-я Царств', 'Четвертая книга Царств [Вторая Царей]'], '4 Царств',
+        BookName(['4 Царств', '4-я Царств', 'Четвертая книга Царств [Вторая Царей]', 'Четвертая книга Царств'],
+                 '4 Царств',
                  '2KI'),
         BookName(['1 Паралипоменон', '1-я Паралипоменон', 'Первая книга Паралипоменон'],
                  '1 Паралипоменон', '1CH'),
@@ -97,39 +99,61 @@ class BookNameManager:
         BookName(['От Луки', 'От Луки святое благовествование'], 'От Луки', 'LUK'),
         BookName(['От Иоанна', 'От Иоанна святое благовествование'], 'От Иоанна', 'JHN'),
         BookName(['Деяния', 'Деяния святых апостолов'], 'Деяния', 'ACT'),
-        BookName(['К Римлянам', 'Послание к Римлянам святого апостола Павла'], 'К Римлянам', 'ROM'),
-        BookName(['1 Коринфянам', '1-е Коринфянам', 'Первое послание к Коринфянам святого апостола Павла'],
+        BookName(['К Римлянам', 'Послание к Римлянам святого апостола Павла', 'Послание к Римлянам'], 'К Римлянам',
+                 'ROM'),
+        BookName(['1 Коринфянам', '1-е Коринфянам', 'Первое послание к Коринфянам святого апостола Павла',
+                  'Первое послание к Коринфянам'],
                  '1 Коринфянам',
                  '1CO'),
-        BookName(['2 Коринфянам', '2-е Коринфянам', 'Второе послание к Коринфянам святого апостола Павла'],
+        BookName(['2 Коринфянам', '2-е Коринфянам', 'Второе послание к Коринфянам святого апостола Павла',
+                  'Второе послание к Коринфянам'],
                  '2 Коринфянам',
                  '2CO'),
-        BookName(['К Галатам', 'Послание к Галатам святого апостоля Павла'], 'К Галатам', 'GAL'),
-        BookName(['К Ефесянам', 'Послание к Ефесянам святого апостола Павла'], 'К Ефесянам', 'EPH'),
-        BookName(['К Филиппийцам', 'Послание к Филиппийцам святого апостола Павла'], 'К Филиппийцам', 'PHP'),
-        BookName(['К Колоссянам', 'Послание к Колоссянам святого апостола Павла'], 'К Колоссянам', 'COL'),
+        BookName(['К Галатам', 'Послание к Галатам святого апостоля Павла', 'Послание к Галатам'], 'К Галатам', 'GAL'),
+        BookName(['К Ефесянам', 'Послание к Ефесянам святого апостола Павла', 'Послание к Ефесянам'], 'К Ефесянам',
+                 'EPH'),
+        BookName(['К Филиппийцам', 'Послание к Филиппийцам святого апостола Павла', 'Послание к Филиппийцам'],
+                 'К Филиппийцам', 'PHP'),
+        BookName(['К Колоссянам', 'Послание к Колоссянам святого апостола Павла', 'Послание к Колоссянам'],
+                 'К Колоссянам', 'COL'),
         BookName(['1 Фессалоникийцам', '1-е Фессалоникийцам',
-                  'Первое послание к Фессалоникийцам (Солунянам) святого апостола Павла'], '1 Фессалоникийцам',
+                  'Первое послание к Фессалоникийцам (Солунянам) святого апостола Павла',
+                  'Первое послание к Фессалоникийцам...', 'Первое послание к Фессалоникийцам'], '1 Фессалоникийцам',
                  '1TH'),
         BookName(['2 Фессалоникийцам', '2-е Фессалоникийцам',
-                  'Второе послание к Фессалоникийцам (Солунянам) святого апостола Павла'], '2 Фессалоникийцам',
+                  'Второе послание к Фессалоникийцам (Солунянам) святого апостола Павла',
+                  'Второе послание к Фессалоникийцам...', 'Второе послание к Фессалоникийцам'], '2 Фессалоникийцам',
                  '2TH'),
-        BookName(['1 Тимофею', '1-е Тимофею', 'Первое послание к Тимофею святого апостола Павла'], '1 Тимофею',
+        BookName(['1 Тимофею', '1-е Тимофею', 'Первое послание к Тимофею святого апостола Павла',
+                  'Первое послание к Тимофею'], '1 Тимофею',
                  '1TI'),
-        BookName(['2 Тимофею', '2-е Тимофею', 'Второе послание к Тимофею святого апостола Павла'], '2 Тимофею',
+        BookName(['2 Тимофею', '2-е Тимофею', 'Второе послание к Тимофею святого апостола Павла',
+                  'Второе послание к Тимофею'], '2 Тимофею',
                  '2TI'),
-        BookName(['К Титу', 'Послание к Титу святого апостола Павла'], 'К Титу', 'TIT'),
-        BookName(['К Филимону', 'Послание к Филимону святого апостола Павла'], 'К Филимону', 'PHM'),
-        BookName(['К Евреям', 'Послание к Евреям святого апостола Павла'], 'К Евреям', 'HEB'),
-        BookName(['Иакова', 'Соборное послание святого апостола Иакова'], 'Иакова', 'JAS'),
-        BookName(['1 Петра', '1-е Петра', 'Первое соборное послание святого апостола Петра'], '1 Петра', '1PE'),
-        BookName(['2 Петра', '2-е Петра', 'Второе соборное послание святого апостола Петра'], '2 Петра', '2PE'),
-        BookName(['1 Иоанна', '1-е Иоанна', 'Первое соборное послание святого апостола Иоанна'], '1 Иоанна',
-                 '1JN'),
-        BookName(['2 Иоанна', '2-е Иоанна', 'Второе соборное послание святого апостола Иоанна'], '2 Иоанна',
-                 '2JN'),
-        BookName(['3 Иоанна', '3-е Иоанна', 'Третье соборное послание святого апостола Иоанна'], '3 Иоанна',
-                 '3JN'),
-        BookName(['Иуды', 'Соборное послание святого апостола Иуды'], 'Иуды', 'JUD'),
-        BookName(['Откровение', 'Откровение святого Иоанна Богослова (Апокалипсис)'], 'Откровение', 'REV')
+        BookName(['К Титу', 'Послание к Титу святого апостола Павла', 'Послание к Титу'], 'К Титу', 'TIT'),
+        BookName(['К Филимону', 'Послание к Филимону святого апостола Павла', 'Послание к Филимону'], 'К Филимону',
+                 'PHM'),
+        BookName(['К Евреям', 'Послание к Евреям святого апостола Павла', 'Послание к Евреям'], 'К Евреям', 'HEB'),
+        BookName(['Иакова', 'Соборное послание святого апостола Иакова', 'Послание Иакова'], 'Иакова', 'JAS'),
+        BookName(['1 Петра', '1-е Петра', 'Первое соборное послание святого апостола Петра', 'Первое послание Петра'],
+                 '1 Петра', '1PE'),
+        BookName(['2 Петра', '2-е Петра', 'Второе соборное послание святого апостола Петра', 'Второе послание Петра'],
+                 '2 Петра', '2PE'),
+        BookName(
+            ['1 Иоанна', '1-е Иоанна', 'Первое соборное послание святого апостола Иоанна', 'Первое послание Иоанна'],
+            '1 Иоанна',
+            '1JN'),
+        BookName(
+            ['2 Иоанна', '2-е Иоанна', 'Второе соборное послание святого апостола Иоанна', 'Второе послание Иоанна'],
+            '2 Иоанна',
+            '2JN'),
+        BookName(
+            ['3 Иоанна', '3-е Иоанна', 'Третье соборное послание святого апостола Иоанна', 'Третье послание Иоанна'],
+            '3 Иоанна',
+            '3JN'),
+        BookName(['Иуды', 'Соборное послание святого апостола Иуды', 'Послание Иуды'], 'Иуды', 'JUD'),
+        BookName(
+            ['Откровение', 'Откровение святого Иоанна Богослова (Апокалипсис)', 'Откровение ап. Иоанна Богослова...',
+             'Откровение ап. Иоанна Богослова'],
+            'Откровение', 'REV')
     ]
